refactor: Table 폴더 변경
