+++ src/app/(auth)/(navigationsBarLayout)/problems/page.tsx
@@ -1,4 +1,4 @@
-"use client";
+'use client';
 
 import {
   Select,
@@ -7,16 +7,17 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select";
-import ProblemTable from "./Table";
-import { useState } from "react";
-import Image from "next/image";
+} from '@/components/ui/select';
+
+import { useState } from 'react';
+import Image from 'next/image';
+import ProblemTable from './_ui/Table';
 
 const ProblemsList = () => {
-  const categoryCodeItem = ["전체", "BFS", "DFS", "수학", "조건문"];
-  const difficultyItem = ["전체", "LV1", "LV2", "LV3", "LV4", "LV5", "LV6", "LV7"];
-  const [categoryCode, setCategoryCode] = useState("");
-  const [difficulty, setDifficulty] = useState("");
+  const categoryCodeItem = ['전체', 'BFS', 'DFS', '수학', '조건문'];
+  const difficultyItem = ['전체', 'LV1', 'LV2', 'LV3', 'LV4', 'LV5', 'LV6', 'LV7'];
+  const [categoryCode, setCategoryCode] = useState('');
+  const [difficulty, setDifficulty] = useState('');
 
   return (
     <div className="flex flex-col px-10 py-18 w-full pt-[140px] gap-4">
@@ -31,7 +32,7 @@ const ProblemsList = () => {
               <SelectTrigger>
                 <SelectValue placeholder="카테고리"></SelectValue>
               </SelectTrigger>
-              <SelectContent style={{ zIndex: 999, position: "relative", backgroundColor: "#000" }}>
+              <SelectContent style={{ zIndex: 999, position: 'relative', backgroundColor: '#000' }}>
                 <SelectGroup>
                   {categoryCodeItem.map((item) => {
                     return (
@@ -52,7 +53,7 @@ const ProblemsList = () => {
               <SelectTrigger>
                 <SelectValue placeholder="난이도"></SelectValue>
               </SelectTrigger>
-              <SelectContent style={{ zIndex: 999, position: "relative", backgroundColor: "#000" }}>
+              <SelectContent style={{ zIndex: 999, position: 'relative', backgroundColor: '#000' }}>
                 <SelectGroup>
                   {difficultyItem.map((item) => {
                     return (
@@ -81,9 +82,9 @@ const ProblemsList = () => {
             </svg>
           </div>
         </div>
-        {(categoryCode !== "전체" || difficulty !== "전체") && (
+        {(categoryCode !== '전체' || difficulty !== '전체') && (
           <div className="flex flex-row gap-8">
-            {categoryCode !== "전체" && categoryCode && (
+            {categoryCode !== '전체' && categoryCode && (
               <div className="flex flex-row gap-1 items-center">
                 <span>{categoryCode}</span>
                 <Image
@@ -93,12 +94,12 @@ const ProblemsList = () => {
                   width={16}
                   height={16}
                   onClick={() => {
-                    setCategoryCode("");
+                    setCategoryCode('');
                   }}
                 />
               </div>
             )}
-            {difficulty !== "전체" && difficulty && (
+            {difficulty !== '전체' && difficulty && (
               <div className="flex flex-row gap-1 items-center">
                 <span>{difficulty}</span>
                 <Image
@@ -108,7 +109,7 @@ const ProblemsList = () => {
                   width={16}
                   height={16}
                   onClick={() => {
-                    setDifficulty("");
+                    setDifficulty('');
                   }}
                 />
               </div>
